requirements.txt
autopep8
Flask==3.0.3
gunicorn==22.0.0
aniso8601==10.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
requests==2.32.3
python-dotenv==1.1.0
psycopg2-binary==2.9.10
Flask-Migrate==4.1.0